Fix Enumeration.TrimmedPaths
